+++ src/main/java/io/github/astrapi69/swing/check/model/JCheckBoxDecorator.java
@@ -24,18 +24,14 @@
  */
 package io.github.astrapi69.swing.check.model;
 
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-
 import javax.swing.*;
 
 import lombok.AccessLevel;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
-import lombok.NonNull;
 import lombok.ToString;
 import lombok.experimental.FieldDefaults;
-import io.github.astrapi69.model.BaseModel;
+import io.github.astrapi69.model.PropertyModel;
 import io.github.astrapi69.model.api.Model;
 
 @Getter
@@ -46,98 +42,46 @@ public class JCheckBoxDecorator extends JCheckBox
 {
 
 	/** The model. */
-	Model<Boolean> bindModel;
+	final Model<Boolean> propertyModel = PropertyModel.<Boolean> of(this, "model.selected");
 
 	public JCheckBoxDecorator()
 	{
-		initializeBindModel(false);
+
 	}
 
 	public JCheckBoxDecorator(Icon icon)
 	{
 		super(icon);
-		initializeBindModel(false);
 	}
 
 	public JCheckBoxDecorator(Icon icon, boolean selected)
 	{
 		super(icon, selected);
-		initializeBindModel(selected);
 	}
 
 	public JCheckBoxDecorator(String text)
 	{
 		super(text);
-		initializeBindModel(false);
 	}
 
 	public JCheckBoxDecorator(Action a)
 	{
 		super(a);
-		initializeBindModel(false);
 	}
 
 	public JCheckBoxDecorator(String text, boolean selected)
 	{
 		super(text, selected);
-		initializeBindModel(selected);
 	}
 
 	public JCheckBoxDecorator(String text, Icon icon)
 	{
 		super(text, icon);
-		initializeBindModel(false);
 	}
 
 	public JCheckBoxDecorator(String text, Icon icon, boolean selected)
 	{
 		super(text, icon, selected);
-		initializeBindModel(selected);
 	}
 
-	@Override
-	public void setModel(ButtonModel newModel)
-	{
-		super.setModel(newModel);
-		if (newModel != null)
-		{
-			if (bindModel == null)
-			{
-				bindModel = BaseModel.of(newModel.isSelected());
-			}
-			else
-			{
-				bindModel.setObject(newModel.isSelected());
-			}
-			newModel.addActionListener(new ActionListener()
-			{
-				@Override
-				public void actionPerformed(ActionEvent e)
-				{
-					bindModel.setObject(newModel.isSelected());
-				}
-			});
-		}
-	}
-
-	public void setBindModel(final @NonNull Model<Boolean> bindModel)
-	{
-		ButtonModel buttonModel = getModel();
-		if (buttonModel == null)
-		{
-			buttonModel = new DefaultButtonModel();
-		}
-		buttonModel.setSelected(bindModel.getObject());
-	}
-
-	public void setBindModelObject(final boolean modelObject)
-	{
-		bindModel.setObject(modelObject);
-		setBindModel(bindModel);
-	}
-
-	protected void initializeBindModel(boolean selected)
-	{
-		bindModel.setObject(selected);
-	}
 }